+++ R/backend-mysql.R
@@ -132,8 +132,8 @@ sql_expr_matches.MySQL <- sql_expr_matches.MariaDBConnection
 sql_expr_matches.MySQLConnection <- sql_expr_matches.MariaDBConnection
 
 #' @export
-sql_values.MariaDBConnection <- function(con, df) {
-  sql_values_clause(con, df, row = TRUE)
+sql_values.MariaDBConnection <- function(con, df, lvl = 0, ...) {
+  sql_values_clause(con, df, row = TRUE, lvl = lvl)
 }
 
 #' @export