+++ R/verb-pivot-wider.R
@@ -230,7 +230,8 @@ dbplyr_pivot_wider_spec <- function(data,
   }
   values_fill <- values_fill[intersect(names(values_fill), values_from_cols)]
 
-  values_fn <- check_list_of_functions(values_fn, values_from_cols, "values_fn")
+  call <- current_env()
+  values_fn <- check_list_of_functions(values_fn, values_from_cols, "values_fn", call = call)
   missing_values <- setdiff(values_from_cols, names(values_fn))
   if (!is_empty(missing_values)) {
     abort("`values_fn` must specify a function for each col in `values_from`")
@@ -238,7 +239,7 @@ dbplyr_pivot_wider_spec <- function(data,
 
   pivot_exprs <- purrr::map(
     set_names(vctrs::vec_seq_along(spec), spec$.name),
-    ~ build_pivot_wider_exprs(.x, spec, values_fill, values_fn, data)
+    ~ build_pivot_wider_exprs(.x, spec, values_fill, values_fn, data, call = call)
   )
 
   key_vars <- setdiff(id_cols, non_id_cols)
@@ -255,7 +256,7 @@ dbplyr_pivot_wider_spec <- function(data,
 
   unused_cols <- setdiff(colnames(data), c(id_cols, non_id_cols))
   unused_fn <- check_list_of_functions(unused_fn, unused_cols, "unused_fn")
-  unused_col_expr <- purrr::imap(unused_fn, ~ resolve_fun(.x, sym(.y), data))
+  unused_col_expr <- purrr::imap(unused_fn, ~ resolve_fun(.x, sym(.y), data, call = call))
 
   data_grouped %>%
     summarise(
@@ -271,25 +272,27 @@ globalVariables(c("name", "value"))
 build_wider_id_cols_expr <- function(data,
                                      id_cols = NULL,
                                      names_from = name,
-                                     values_from = value) {
+                                     values_from = value,
+                                     call = caller_env()) {
   # COPIED FROM tidyr
   # TODO: Use `allow_rename = FALSE`.
   # Requires https://github.com/r-lib/tidyselect/issues/225.
   sim_data <- simulate_vars(data)
-  names_from <- names(tidyselect::eval_select(enquo(names_from), sim_data))
-  values_from <- names(tidyselect::eval_select(enquo(values_from), sim_data))
+  names_from <- names(tidyselect::eval_select(enquo(names_from), sim_data, error_call = call))
+  values_from <- names(tidyselect::eval_select(enquo(values_from), sim_data, error_call = call))
   non_id_cols <- c(names_from, values_from)
 
   out <- select_wider_id_cols(
     data = data,
     id_cols = {{id_cols}},
-    non_id_cols = non_id_cols
+    non_id_cols = non_id_cols,
+    call = call
   )
 
   expr(c(!!!out))
 }
 
-build_pivot_wider_exprs <- function(row_id, spec, values_fill, values_fn, data) {
+build_pivot_wider_exprs <- function(row_id, spec, values_fill, values_fn, data, call) {
   values_col <- spec[[".value"]][row_id]
   fill_value <- values_fill[[values_col]]
 
@@ -305,12 +308,13 @@ build_pivot_wider_exprs <- function(row_id, spec, values_fill, values_fn, data) 
   case_expr <- expr(ifelse(!!keys_cond, !!sym(values_col), !!fill_value))
 
   agg_fn <- values_fn[[values_col]]
-  resolve_fun(agg_fn, case_expr, data = data)
+  resolve_fun(agg_fn, case_expr, data = data, call = call)
 }
 
 select_wider_id_cols <- function(data,
                                  id_cols = NULL,
-                                 non_id_cols = character()) {
+                                 non_id_cols = character(),
+                                 call = caller_env()) {
   # COPIED FROM tidyr
   id_cols <- enquo(id_cols)
   sim_data <- simulate_vars(data)
@@ -323,7 +327,7 @@ select_wider_id_cols <- function(data,
   } else {
     # TODO: Use `allow_rename = FALSE`.
     # Requires https://github.com/r-lib/tidyselect/issues/225.
-    names(tidyselect::eval_select(enquo(id_cols), sim_data))
+    names(tidyselect::eval_select(enquo(id_cols), sim_data, error_call = call))
   }
 }
 
@@ -340,14 +344,14 @@ is_scalar <- function(x) {
 }
 
 
-resolve_fun <- function(x, var, data) {
+resolve_fun <- function(x, var, data, call = caller_env()) {
   if (is_formula(x)) {
     .fn_expr <- across_fun(x, env = empty_env(), data = data, dots = NULL, fn = "across")
     exec(.fn_expr, var)
   } else {
     fn_name <- find_fun(x)
     if (is_null(fn_name)) {
-      abort("Can't convert to a function.")
+      abort("Can't convert to a function.", call = call)
     }
     call2(fn_name, var)
   }