+++ NEWS.md
@@ -3,6 +3,13 @@
 * Support `rows_update()`, `rows_patch()`, `rows_upsert()`, and `rows_delete()`
   (@mgirlich, #736).
 
+* Added translations for lubridate functions `day()`, `week()`, `isoweek()`,
+  and `isoyear()` for Postgres (@mgirlich, #675).
+
+* Fix `union()` translation for Hive (@mgirlich, #663).
+
+* Fix translation of `as.Date()` for Oracle (@mgirlich, #661).
+
 * Added translation for `cut()` (@mgirlich, #697).
 
 * Fix translation of `str_flatten()` for Redshift (@hdplsa, #804)